review: finish remaining hardening fixes
Align protocol tab indices with the FlatBuffers u32 contract across the CLI, client, server, and snapshot decoding paths, and keep PTY helper threads owned by the buffer runtime so shutdown and exit observe thread failures instead of dropping them silently.

+++ crates/embers-client/src/scripting/context.rs
@@ -249,7 +249,7 @@ pub struct NodeRef {
     pub child_ids: Vec<NodeId>,
     pub geometry: Option<Rect>,
     pub tab_titles: Vec<String>,
-    pub active_tab: Option<usize>,
+    pub active_tab: Option<u32>,
     pub buffer_id: Option<BufferId>,
     pub visible: bool,
 }
